Add GetSchemaFieldAPI class
