+++ solr/core/src/java/org/apache/solr/jersey/InjectionFactories.java
@@ -18,9 +18,12 @@
 package org.apache.solr.jersey;
 
 import static org.apache.solr.jersey.RequestContextKeys.SOLR_CORE;
+import static org.apache.solr.jersey.RequestContextKeys.SOLR_PARAMS;
 
 import javax.inject.Inject;
 import javax.ws.rs.container.ContainerRequestContext;
+
+import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.core.SolrCore;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.response.SolrQueryResponse;
@@ -102,6 +105,24 @@ public class InjectionFactories {
 
     @Override
     public void dispose(IndexSchema instance) {}
+  }
+
+  public static class ReuseFromContextSolrParamsFactory implements Factory<SolrParams> {
+
+    private final ContainerRequestContext containerRequestContext;
+
+    @Inject
+    public ReuseFromContextSolrParamsFactory(ContainerRequestContext containerRequestContext) {
+      this.containerRequestContext = containerRequestContext;
+    }
+
+    @Override
+    public SolrParams provide() {
+      return (SolrParams) containerRequestContext.getProperty(SOLR_PARAMS);
+    }
+
+    @Override
+    public void dispose(SolrParams instance) {}
   }
 
   public static class SingletonFactory<T> implements Factory<T> {
