+++ solr/core/src/java/org/apache/solr/jersey/JerseyApplications.java
@@ -21,6 +21,8 @@ import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Info;
 import io.swagger.v3.oas.annotations.info.License;
 import java.util.Map;
+
+import org.apache.solr.common.params.SolrParams;
 import org.apache.solr.core.SolrCore;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.response.SolrQueryResponse;
@@ -120,6 +122,9 @@ public class JerseyApplications {
               bindFactory(InjectionFactories.ReuseFromContextIndexSchemaFactory.class)
                   .to(IndexSchema.class)
                   .in(RequestScoped.class);
+              bindFactory(InjectionFactories.ReuseFromContextSolrParamsFactory.class)
+                .to(SolrParams.class)
+                .in(RequestScoped.class);
             }
           });
     }
